DON-1855 cleanup on blur
