+++ Backpack-SwiftUI/Blur/Classes/BPKBlur.swift
@@ -25,14 +25,9 @@ public struct BPKBlur {
     /// The radius of the blur effect.
     let radius: CGFloat
 
-    /// Whether the blur should treat the view as opaque.
-    /// Setting this to `true` can improve performance if transparency isn't needed.
-    let opaque: Bool = true
-
     /// Creates a new instance of `BPKBlur`.
     /// - Parameters:
     ///   - radius: The blur radius to apply.
-    ///   - opaque: Whether the blur should be opaque (`true`) or allow transparency (`false`). Defaults to `false`.`.
     public init(radius: CGFloat) {
         self.radius = radius
     }
@@ -43,6 +38,6 @@ public extension View {
     /// - Parameter blur: A `BPKBlur` configuration defining the type and style of the blur.
     /// - Returns: A modified view with the blur applied.
     func blur(_ blur: BPKBlur) -> some View {
-        self.blur(radius: blur.radius, opaque: blur.opaque)
+        self.blur(radius: blur.radius, opaque: true)
     }
 }
